requirements.txt
absl-py==2.3.1
alembic==1.17.2
anyio==4.12.0
astunparse==1.6.3
bcrypt==5.0.0
beautifulsoup4==4.14.2
blinker==1.9.0
CacheControl==0.14.4
cachetools==6.2.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
coloredlogs==15.0.1
cryptography==46.0.3
filelock==3.20.0
firebase_admin==7.1.0
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
flatbuffers==25.9.23
gast==0.6.0
gdown==5.2.0
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.6.0
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.76.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
h5py==3.15.1
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
libclang==18.1.1
Mako==1.3.10
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
msgpack==1.1.2
namex==0.1.0
nltk==3.9.2
numpy==2.2.6
onnx==1.17.0
onnxruntime==1.23.2
opencv-python-headless==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
proto-plus==1.26.1
protobuf==6.33.1
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
skl2onnx==1.19.1
soupsieve==2.8
SQLAlchemy==2.0.44
sympy==1.14.0
termcolor==3.2.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.1
